refactor: SejongAuthenticator 개선 [#38]
